% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runArrayAnalysis.R
\name{runArrayAnalysis}
\alias{runArrayAnalysis}
\title{runArrayAnalysis}
\usage{
runArrayAnalysis()
}
\value{
This function will run the ArrayAnalysis app
}
\description{
runArrayAnalysis
}
\examples{
runArrayAnalysis()
}
